feat: split tests of `query.max_query_count`
