+++ tests/DependencyInjection/ConfigurationConfigMaxQueryCountTest.php
@@ -11,12 +11,13 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Liip\Acme\Tests\DependencyInjection;
+namespace DependencyInjection;
 
-use Liip\Acme\Tests\AppConfig\AppConfigKernel;
+use Liip\Acme\Tests\AppConfigMaxQueryCount\AppConfigMaxQueryCountKernel;
+use Liip\Acme\Tests\DependencyInjection\ConfigurationTest;
 
 /**
- * Use Tests/AppConfig/AppConfigKernel.php instead of
+ * Use Tests/AppConfigMaxQueryCount/AppConfigMaxQueryCountKernel.php instead of
  * Tests/App/AppKernel.php.
  * So it must be loaded in a separate process.
  *
@@ -24,14 +25,14 @@ use Liip\Acme\Tests\AppConfig\AppConfigKernel;
  *
  * @preserveGlobalState disabled
  */
-class ConfigurationConfigTest extends ConfigurationTest
+class ConfigurationConfigMaxQueryCountTest extends ConfigurationTest
 {
     /**
      * Use another Kernel to load another config file.
      */
     protected static function getKernelClass(): string
     {
-        return AppConfigKernel::class;
+        return AppConfigMaxQueryCountKernel::class;
     }
 
     /**
@@ -42,7 +43,7 @@ class ConfigurationConfigTest extends ConfigurationTest
         return [
             ['command_verbosity', 'very_verbose'],
             ['command_decoration', false],
-            ['query.max_query_count', 1000],
+            ['query.max_query_count', 0],
             ['authentication.username', 'foobar'],
             ['authentication.password', '12341234'],
         ];
