+++ tests/Test/WebTestCaseConfigTest.php
@@ -13,8 +13,6 @@ declare(strict_types=1);
 
 namespace Liip\Acme\Tests\Test;
 
-use Doctrine\Common\Annotations\Annotation\IgnoreAnnotation;
-use Liip\Acme\Tests\App\Entity\User;
 use Liip\Acme\Tests\AppConfig\AppConfigKernel;
 use Liip\Acme\Tests\Traits\LiipAcmeFixturesTrait;
 use Liip\FunctionalTestBundle\Annotations\QueryCount;
@@ -31,11 +29,6 @@ use Symfony\Component\HttpKernel\Kernel;
  * @runTestsInSeparateProcesses
  *
  * @preserveGlobalState disabled
- *
- * Avoid conflict with PHPUnit annotation when reading QueryCount
- * annotation:
- *
- * @IgnoreAnnotation("expectedException")
  */
 class WebTestCaseConfigTest extends WebTestCase
 {
@@ -43,6 +36,13 @@ class WebTestCaseConfigTest extends WebTestCase
 
     /** @var \Symfony\Bundle\FrameworkBundle\Client client */
     private $client;
+
+    protected function tearDown(): void
+    {
+        parent::tearDown();
+
+        restore_exception_handler();
+    }
 
     protected static function getKernelClass(): string
     {
@@ -54,8 +54,6 @@ class WebTestCaseConfigTest extends WebTestCase
      */
     public function testIndexClientWithCredentials(): void
     {
-        $this->skipTestIfSymfonyHasVersion7();
-
         $this->client = static::makeClientWithCredentials('foobar', '12341234');
 
         $path = '/admin';
@@ -87,8 +85,6 @@ class WebTestCaseConfigTest extends WebTestCase
      */
     public function testIndexAuthenticatedClient(): void
     {
-        $this->skipTestIfSymfonyHasVersion7();
-
         $this->client = static::makeAuthenticatedClient();
 
         $path = '/admin';
@@ -120,8 +116,6 @@ class WebTestCaseConfigTest extends WebTestCase
      */
     public function testIndexAuthenticationLoginAs(): void
     {
-        $this->skipTestIfSymfonyHasVersion7();
-
         $user = $this->loadTestFixtures();
 
         $loginAs = $this->loginAs($user, 'secured_area');
@@ -160,8 +154,6 @@ class WebTestCaseConfigTest extends WebTestCase
      */
     public function testIndexAuthenticationLoginClient(): void
     {
-        $this->skipTestIfSymfonyHasVersion7();
-
         $user = $this->loadTestFixtures();
 
         $this->client = static::makeClient();
@@ -187,57 +179,6 @@ class WebTestCaseConfigTest extends WebTestCase
         $this->assertSame(
             'LiipFunctionalTestBundle',
             $crawler->filter('h1')->text()
-        );
-    }
-
-    /**
-     * Log in as the user defined in the Data Fixtures and except an
-     * AllowedQueriesExceededException exception.
-     *
-     * There will be 2 queries:
-     * - the user 1 is loaded from the database when logging in
-     * - the user 2 is loaded by the controller
-     *
-     * In the configuration the limit is 1, an Exception will be thrown.
-     */
-    public function testAllowedQueriesExceededException(): void
-    {
-        $this->skipTestIfSymfonyHasVersion7();
-
-        $user = $this->loadTestFixtures();
-
-        $this->assertInstanceOf(
-            User::class,
-            $user
-        );
-
-        $this->client = static::makeClient();
-
-        $this->loginClient($this->client, $user, 'secured_area');
-
-        $path = '/user/2';
-
-        $this->expectException(\Liip\FunctionalTestBundle\Exception\AllowedQueriesExceededException::class);
-
-        $crawler = $this->client->request('GET', $path);
-
-        // The following code is called if no exception has been thrown, it should help to understand why
-        $this->assertStatusCode(200, $this->client);
-        $this->assertSame(
-            'LiipFunctionalTestBundle',
-            $crawler->filter('h1')->text()
-        );
-        $this->assertSame(
-            'Logged in as foo bar.',
-            $crawler->filter('p#user')->text()
-        );
-        $this->assertSame(
-            'Name: alice bob',
-            $crawler->filter('div#content p:nth-child(1)')->text()
-        );
-        $this->assertSame(
-            'Email: alice@example.com',
-            $crawler->filter('div#content p:nth-child(2)')->text()
         );
     }
